backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
supabase==2.3.4
python-dotenv==1.0.1
httpx==0.27.0
